+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   Mailto links in error alerts (#572)
+-   Download buttons on ComputePlans page (#581)
 
 ### Fixed